requirements.txt
gTTS==2.3.1
pygame==2.2.0
python-dotenv==1.0.0
twitchio==2.5.0
